speech-diff/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
aniso8601==9.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==19.10b0
boto3==1.28.44
botocore==1.31.44
braceexpand==0.1.7
cachetools==5.3.1
cdifflib==1.2.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
cytoolz==0.12.2
dataclasses==0.6
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastdtw==0.3.4
fasttext==0.9.2
filelock==3.12.3
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
ijson==3.2.3
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
ipadic==1.0.0
ipython==8.15.0
ipywidgets==8.1.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab-widgets==3.0.8
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.0
latexcodec==2.0.1
lazy_loader==0.3
Levenshtein==0.21.1
lhotse==1.16.0
librosa==0.10.1
lightning-utilities==0.9.0
lilcom==1.7
lit==16.0.6
llvmlite==0.40.1
loguru==0.7.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.5
megatron-core==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-text-processing==0.2.0rc0
nemo-toolkit==1.20.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
OpenCC==1.1.6
packaging==23.1
pandas==2.1.0
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
plac==1.3.5
platformdirs==3.10.0
pluggy==1.3.0
pooch==1.7.0
portalocker==2.7.0
progress==1.6
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.16.1
pynini==2.1.5
pyparsing==3.0.9
pypinyin==0.49.0
pypinyin-dict==0.6.0
PySocks==1.7.1
pysptk==0.2.1
pytest==7.4.2
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2023.3.post1
pyworld==0.3.4
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
shellingham==1.5.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.6
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-bibtex==2.6.1
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
text-unidecode==1.3
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.1.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
triton==2.0.0
typed-ast==1.5.5
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.16
wandb==0.15.10
wcwidth==0.2.6
webdataset==0.1.62
Werkzeug==2.3.7
wget==3.2
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.9.2
youtokentome==1.0.6
zipp==3.16.2
